requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
Django==3.2.1
django-admin-honeypot-updated-2021==1.2.0
django-bootstrap-v5==1.0.11
django-ipware==5.0.0
django-jsonfield==1.4.1
pytz==2023.3
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
typing-extensions==4.7.1
